ACPI Shutdown!
Finished ACPI-controlled shutdown, with a command and all! Also added emulator-based shutdowns. Orchid now senses when an emulator is running it as well. Added more BOOT_ERROR_FLAGS. Added a KNOWN ISSUES section to the parser. It is a flawed piece that needs to be cleaned with the keyboard IRQ and the screen wrapper functions.