Minor RAM Fixes & Ethernet Groundwork
Laying the groundwork for Ethernet implementation in a virtual machine. Once this is done, I will be adding separate adapters based on
brandname and how common they are. RAM is now represented in MiB in the SYS command call. The system was rejecting a Heap of >2GiB due to a
bad signing error, which is now fixed. The system should now support up to 4GiB, even if the computer physically has more than that.