Laying Groundwork for VFS in RAM & Introducing BLOOM
Starting to lay the groundwork for a VFS in system RAM, and thinking about maintaining the 16 MiB Heap, or rather moving it on top of the VFS so it can grow. My only fear about the Ephemeral OS concept is that the systems that run the OS may not have the RAM for the applications/files they would like to load. Address this later by introducing a minimum RAM requirement. Introducing BLOOM, orchid's pre-compile byte-code configuration language that will allow users to write their own programs and "install" them onto their systems, maintaining a config even though the OS is still ephemeral in nature.