Added CHANGELOG & Working With VFS
Added a CHANGELOG file to better track current and future version histories with the Orchid platform. Began real implementation of the VFS layout in RAM; placed it in the stead of the Heap, and threw the Heap on top of it so it can still expand. Added STRSCAN, a very useful utility to find a matching string in memory (verified in the snippets.asm file with working code).