Strlen, Processes, & Fixed DUMP Command
I wasn't even aware that DUMP was broken, how embarrassing. :( Added some more OPS functions (strlen) and light process management for the Heap. Beginning to allocate RX and TX buffers for the Ethernet driver, so it will be ready to R/TX frames with payloads when I implement it. COMING SOON: Ethernet connectivity, a new process management command (!), and more with the VFS to handle incoming data accordingly. I need to increase its size! Thinking of making Orchid x86 require a minimum of 128 MiB, which is certainly reasonable.