Created System Font, Tinkering with GUI_MODE
Added a system Monospace font (almost 2000 lines with functions!), which turned out looking just a liiiiiittle quirky when written out, but I was amazed I got it to display as quickly and as easily as I did. This doesn't really reflect what I should be working on (based on the changelog) but the break to do this was quite nice. Working more on a more vital component next.