Registered Ethernet PIC IRQ42
The first (and only the first) Ethernet device found on the PCI bus is now assigned IRQ 42 for hardware interruptions. Fixed my PCI write function FINALLY (the entire file needs to be split properly, and renovated), and spent literally a whole day's work figuring it all out. :) Added a Heap tidbit so that the allocation in KMALLOC will now point the processors in front of the HEAP_HEADER, rather than pointing it to the base of the header. Playing with qemu configs in the Windows compiler to start getting virtual network packet captures. Tomorrow should be the beginning of an amazing journey into Layer 1 (Physical) of the OSI Model. Looking forward to ARP requests. :+1: