Cleaning, Macros, & Ethernet RX/TX - Part 2/?
This hurts my brain. All of the registers show the appropriate values according to the I217 Intel adapter's specifications for normal operation, but the device just isn't functioning, neither TX initalizations or RX IRQs. I have to be missing a step. Also introduced an AMAZING readability feature of the source. You'll start to see more and more parts shrink into more digestable lengths as macros like the new func() one are introduced. I wanted to commit to GitHub now because IF, for some reason, these macros destroy any foundation of the work, I must be able to have a clear point of introduction to reference.