Very Close to Working MD5
The Message Digest algorithm is a slightly complex task in Assembly. Though I've nailed down the majority of the required steps (padding, checking, etc) in the process, something is still causing the computations to churn out the wrong results. You'll see in the commented sections that I've configured a few different methods of computation, none of which have yet worked. I believe my current code is the closest I've gotten to success. I will figure this out eventually (I hope), for the sake of sanity and peace of mind. This algorithm is going to be vital in verifying RX/TX data integrity.