Cleanup Sweep
Documentation adjustments. Adding the memory-zeroing function when shutting down/rebooting; adding more zones to it later. Fixed the system not rebooting when using the command. Splitting up some of the large files (still have to do PCI). More VFS stuff (more coming later). Cleaning up the source code for readability. A lot of consolidation and deletions of redundant things. There's still much to do here, but I will continue it at a later time when the entropy starts to get to me again.