Fixing Issue #3 and Some Scaffolding
Fixed the COMMAND_DUMP issue (see Issue #3) and created some more scaffolding for GUI_MODE. I know that the next part in the growth of the project may cause an explosion in the compilation size, due to the need for so many video variables. I am leery about developing this right now because of that, but I'm excited to potentially get out of SHELL_MODE and into a better environment.