Scaffolding pt. 3/?
Putting together some more scaffolding for the work I want to do in the future, but now it just feels like I'm stalling for time. Still not able to get MD5 to work, unfortunately. May need a fresh mind to view the code and give me a sanity check. Version 0.4 is almost finished.