Readying New Project Branch
Readying a new starting point for a branch in the repo that will create a SYSCALL function and likely result in a complete change to the current trajectory. Lots of entropy starting in the project and I'm looking to take the web diagram in the README, and turn it into a top-down flowchart of OS libraries/functions/zones instead. No more isolated components that are communicating through obscure inter-kernel function calls. I will continue to work on the repo as the CHANGELOG suggests, with the branch being the primary focus however.